requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
SQLAlchemy==2.0.0
pytest==7.2.1
asyncpg==0.27.0
sqlmodel==0.0.8
python_jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5